Update download.py and update.py
The update and download scripts have been updated
to search for the script tag inside the video iframe.
After the iframe is found then regex is used to find the
video url.

The original method of finding the URL was failing to
match any URLs.The VPN message has also been updated
to match on channels that require a VPN.

Updated several xpath and iframe selenium calls to
clean up the deprecation warnings.

Fixed the regex to be more defined for finding the URL in
the iframe.